refactor: strategic pattern package move
