+++ src/main/java/racingcar/model/CarName.java
@@ -1,0 +1,2 @@
+package racingcar.model;public class CarName {
+}
