+++ src/main/java/racingcar/strategy/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.strategy;
 
 import racingcar.util.RandomRange;
